vdev Auction

Functions

edit_item(id)

get_bids_for_user(user)

get_item(id)

get_item_by(attrs)

get_item_with_bids(id)

get_user(id)

get_user_by_username_and_password(username, password)

Retrives a User from the database matching the provided username and password

Return values

Depending on what is found in the database, two different values couldbe returned:

an Auction.User struct: An Auction.User record was found that matched the username and password that was provided.

false: No Auction.User could be found with the provided username and password.

You can then use the returned value to determine whether or not the User is authorized in your application. If an Auction.User is not found based on username, the computational work of hashing a password is still done.

insert_bid(params)

insert_item(attrs)

insert_user(params)

list_items()

new_bid()

new_item()

new_user()

update_item(item, updates)